+++ doc/luadoc/widget.image.lua
@@ -12,8 +12,9 @@
 
 --- @method filename
 -- Show an image from a file. On HiDPI screens, a file with "@2x" before
--- the extension will be used instead, if it is present. Relative paths
--- will be relative to the luakit installation directory.
+-- the extension will be used instead, if it is present.
+--
+-- The filename is used to search the directories specified by @ref{luakit/resource_path}.
 -- @tparam string path The path to the image file.
 
 --- @method icon